完成日志类型模块接口开发，修改详细设计文档

+++ src/main/java/chenweipan/journal/web/response/PageResp.java
@@ -3,17 +3,17 @@ package chenweipan.journal.web.response;
 import java.util.List;
 
 public class PageResp<T> {
-    private int total;
+    private Long total;
     private int items;
     private int page;
     private int totalPages;
     private List<T> content;
 
-    public int getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
